parallel collection and fixes

+++ actor_training.py
@@ -1,9 +1,12 @@
 import os
 import shutil
+import sys
 import time
 from collections import defaultdict
 from pathlib import Path
 from typing import Optional
+from multiprocessing import Process, Event, Pipe
+from threading import Thread
 
 import numpy as np
 import torch
@@ -101,6 +104,60 @@ def lambda_return_schedule(epoch, last_epoch=MAX_EPISODES // 2, start_lambda=0.9
     return last_lambda
 
 
+def process_collection(event: Event, config, env: GymEnv,
+                       actor_critic, dataset: EpisodesDataset,
+                       weights_path, log_connection, verbose=10):
+    sys.stdin = open(0)  # otherwise input() will fail
+    ac_trainer = ACTrainer(config, env, actor_critic, None, dataset)
+    while True:
+        try:
+            if event.is_set():
+                # collection
+                actor_weights = torch.load(weights_path, map_location=device)
+                field_name = "actor_critic."
+                actor_weights = {k[len(field_name):]: v for k, v in actor_weights.items() if field_name in k}
+                ac_trainer.actor.load_state_dict(actor_weights)
+
+                for _ in range(config.training.actor_critic.collection_steps):
+                    ac_trainer.reset(collection=True)
+                    skip_update = False
+                    while not env.done:
+                        step_metrics = ac_trainer.step(collection=True)
+                        if verbose and ac_trainer.episode_step % verbose == 0:
+                            print(f"step value {ac_trainer.outputs[-1].means_values[0].item()}")
+                            print(f"actions: {ac_trainer.actions[-1][0].cpu().detach().numpy()}")
+                            print(f"act_logits: {ac_trainer.outputs[-1].logits_actions[0].cpu().detach().numpy()}")
+                            print(f"means_stds: {ac_trainer.outputs[-1].mean_continuous[0].cpu().detach().numpy()}"
+                                  f" {ac_trainer.outputs[-1].std_continuous[0].cpu().detach().numpy()}")
+
+                        if len(ac_trainer.observations) > 20 and \
+                                not np.count_nonzero(ac_trainer.observations[-20] -
+                                                     ac_trainer.observations[-1]) or step_metrics['buggy_episode']:
+                            print("Train loop: env froze, attempting reload")
+                            skip_update = True
+                            break  # if froze inside episode
+
+                    if skip_update:
+                        env.__exit__(soft=False)
+                        RELOAD_MACRO.play()
+                    else:
+                        if not ac_trainer.actions:
+                            print("Empty episode. Skipping")
+                            return
+                        ac_trainer.episode_end(collection=True, do_step=False)
+                        dataset.update_disk_checkpoint(Path("checkpoints/dataset"))
+                    log_connection.send(ac_trainer.metrics)
+
+                # reset event flag
+                event.clear()
+            else:
+                time.sleep(1)
+        except Exception as e:
+            import traceback
+            traceback.print_exc()
+            time.sleep(1)
+
+
 class ACTrainer:
     def __init__(self,
                  cfg,
@@ -113,17 +170,16 @@ class ACTrainer:
         self.env = env
         self.actor = actor.to(device)
         self.optimizer = optimizer
-        self.lr_scheduler = torch.optim.lr_scheduler.LambdaLR(
-            self.optimizer,
-            lambda epoch: warmup_lr_lambda(epoch,
-                                           self.cfg.training.actor_critic.lr_warmup_scale,
-                                           self.cfg.training.actor_critic.lr_warmup_steps)
-        )
-        # self.lr_scheduler = torch.optim.lr_scheduler.OneCycleLR(
-        #     self.optimizer,
-        #     max_lr=self.cfg.training.learning_rate,
-        #     total_steps=MAX_EPISODES
-        # )
+        if self.optimizer is not None:
+            self.lr_scheduler = torch.optim.lr_scheduler.LambdaLR(
+                self.optimizer,
+                lambda epoch: warmup_lr_lambda(epoch,
+                                               self.cfg.training.actor_critic.lr_warmup_scale,
+                                               self.cfg.training.actor_critic.lr_warmup_steps)
+            )
+        else:
+            self.lr_scheduler = None
+
         self.dataset = dataset
         self.replay_only = replay_only
 
@@ -132,8 +188,7 @@ class ACTrainer:
         self.metrics = defaultdict(float)
 
         self.batch_size = self.cfg.training.actor_critic.batch_num_samples
-        self.accumulation_steps = max(self.cfg.training.actor_critic.grad_acc_steps,
-                                      self.cfg.training.actor_critic.collection_steps)
+        self.accumulation_steps = self.cfg.training.actor_critic.grad_acc_steps
 
         self.episode_step = 0
         self._replay = None
@@ -187,6 +242,7 @@ class ACTrainer:
         print("ACTrainer.reset: Started reset")
         self.metrics = defaultdict(float)
         self.episode_step = 0
+        self.dataset.load_disk_checkpoint(Path("checkpoints/dataset"))
 
         # sample episodes via prioritized sampling
         if len(self.dataset) >= self.batch_size - 1 and self.batch_size > 1:
@@ -367,7 +423,9 @@ class ACTrainer:
             if mean_line:
                 ax.axvline(data_[:, i].mean(), color='k', linestyle='dashed', linewidth=1)
         ax.legend()
-        return wandb.Image(f)
+        img = wandb.Image(f)
+        plt.close("all")
+        return img
 
     def episode_end(self, epoch=None, collection=False, do_step=True):
         # finish replaying if needed
@@ -396,7 +454,7 @@ class ACTrainer:
             # try to learn neg reward
             true_reward = self.rewards[self.episode_step - 1][0].item()
             self.rewards[self.episode_step - 1][0] = true_reward if true_reward > 0.375 else (
-                                                                                                            true_reward - 0.375) / 1.375  # positive starts at rank 3
+                                                                                                     true_reward - 0.375) / 1.375  # positive starts at rank 3
             # update value estimation
             self.rewards[self.episode_step - 3][0] = self.rewards[self.episode_step - 1][
                 0]  # move reward closer bc of lambda returns
@@ -410,10 +468,7 @@ class ACTrainer:
             self.dataset.add_episode(
                 self._create_episode()
             )
-
-        if collection:
-            return self.metrics
-
+        actual_batch_size = len(self.dones[-1])
         episode_output = ImagineOutput(
             observations=None,
             actions=torch.stack(self.actions, dim=1)[..., :-3].to(device),
@@ -421,15 +476,14 @@ class ACTrainer:
             logits_actions=torch.cat([out.logits_actions for out in self.outputs], dim=1).to(device),
             continuous_means=torch.cat([out.mean_continuous for out in self.outputs], dim=1).to(device),
             continuous_stds=torch.cat([out.std_continuous for out in self.outputs], dim=1).to(device),
-            values=torch.stack([out.means_values for out in self.outputs], dim=1).reshape(self.batch_size, -1).to(
+            values=torch.stack([out.means_values for out in self.outputs], dim=1).reshape(actual_batch_size, -1).to(
                 device),
-            rewards=torch.stack(self.rewards, dim=1).reshape(self.batch_size, -1).to(device),
-            ends=torch.stack(self.dones, dim=1).reshape(self.batch_size, -1).to(device)
+            rewards=torch.stack(self.rewards, dim=1).reshape(actual_batch_size, -1).to(device),
+            ends=torch.stack(self.dones, dim=1).reshape(actual_batch_size, -1).to(device)
         )
         mask_paddings = torch.stack(self.mask_paddings, dim=1).to(device)
-
-        # log action and value histograms
-        if do_step:  # speed up
+        if collection:
+            # log action and value histograms
             self.metrics["values"] = self._create_hist(episode_output.values[mask_paddings], names=["value"],
                                                        mean_line=True)
             self.metrics["binary_logits"] = self._create_hist(
@@ -438,8 +492,11 @@ class ACTrainer:
                 episode_output.logits_actions[mask_paddings][..., 4:8],
                 episode_output.logits_actions[mask_paddings][..., 8:],
             ], dim=-1), names=ACTION_NAMES[4:])
-            self.metrics["means"] = self._create_hist(episode_output.continuous_means[mask_paddings], names=ACTION_C_NAMES)
-            self.metrics["stds"] = self._create_hist(episode_output.continuous_stds[mask_paddings], names=ACTION_C_NAMES)
+            self.metrics["means"] = self._create_hist(episode_output.continuous_means[mask_paddings],
+                                                      names=ACTION_C_NAMES)
+            self.metrics["stds"] = self._create_hist(episode_output.continuous_stds[mask_paddings],
+                                                     names=ACTION_C_NAMES)
+            return self.metrics
 
         self.metrics["actor_critic/train/lr"] = self.lr_scheduler.get_last_lr()[0]
 
@@ -455,7 +512,7 @@ class ACTrainer:
             self.optimizer.step()
             self.lr_scheduler.step()
             self.optimizer.zero_grad()
-
+        loss = loss / (1 / self.accumulation_steps)
         print("backward update elapsed: ", time.time() - start_backward)
         self._update_metrics(loss.intermediate_losses)
         self.metrics["actor_critic/train/total_loss"] = loss.loss_total.item()
@@ -464,10 +521,11 @@ class ACTrainer:
     def ac_loss(self, outputs: ImagineOutput, mask_paddings, epoch=None):
         mask_paddings = torch.logical_and(mask_paddings, outputs.ends.logical_not())  # do not include end into loss
         values = outputs.values
-        # TODO test value normalization
-        # values_std = torch.masked_select(values, mask_paddings).std().detach()
-        # values_mean = torch.masked_select(values, mask_paddings).mean().detach()
-        # values = (values - values_mean) / values_std
+        values_std = torch.masked_select(values, mask_paddings).std()
+        values_mean = torch.masked_select(values, mask_paddings).mean()
+        self.metrics["value_mean"] = values_mean.item()
+        self.metrics["value_std"] = values_std.item()
+        values = (values - values_mean) / values_std
         if epoch is None or not self.cfg.training.actor_critic.lambda_warmup_init:
             lambda_ = self.cfg.training.actor_critic.lambda_
         else:
@@ -494,10 +552,10 @@ class ACTrainer:
         # self.metrics["advantages"] = self._create_hist(advantage_factor[mask_paddings], mean_line=True)
         # self.metrics["weighted_advantages"] = self._create_hist((update_weight * advantage_factor)[mask_paddings],
         #                                                         mean_line=True)
-        self.metrics["value_mean"] = values[mask_paddings].mean().item()
+        self.metrics["advantage_mean"] = advantage_factor[mask_paddings].mean().item()
         advantage_factor[advantage_factor < 0] /= POSITIVE_WEIGHT
 
-        # compute losses
+        # compute losses  # todo: log -ln(x) | -ln(1-x) where adv > 0
         loss_actions = -1 * (log_probs * advantage_factor.unsqueeze(-1))[..., ACTION_LOCK_MASK]
         loss_actions_masked = torch.masked_select(update_weight.unsqueeze(-1) * loss_actions,
                                                   mask_paddings.unsqueeze(-1)).mean()
@@ -510,10 +568,12 @@ class ACTrainer:
         loss_entropy = torch.masked_select(
             - self.cfg.training.actor_critic.entropy_weight * torch.log(entropy[..., ACTION_LOCK_MASK]),
             mask_paddings.unsqueeze(-1)).mean() + \
-            torch.masked_select(  # logit l2 regularization loss. cat(logits=[10,10,-10,-10]).entropy() is 0.69!!
-                self.cfg.training.actor_critic.entropy_weight * outputs.logits_actions[..., ACTION_LOGIT_LOCK_MASK]**2,
-            mask_paddings.unsqueeze(-1)
-            ).mean()
+                       torch.masked_select(
+                           # logit l2 regularization loss. cat(logits=[10,10,-10,-10]).entropy() is 0.69!!
+                           self.cfg.training.actor_critic.entropy_weight * outputs.logits_actions[
+                               ..., ACTION_LOGIT_LOCK_MASK] ** 2,
+                           mask_paddings.unsqueeze(-1)
+                       ).mean()
         loss_entropy_continuous = torch.masked_select(
             - self.cfg.training.actor_critic.entropy_continuous_weight * entropy_cont[..., ACTION_C_LOCK_MASK],
             mask_paddings.unsqueeze(-1)).mean()
@@ -549,12 +609,12 @@ def main(cfg: DictConfig):
             r"C:\Users\Michael\PycharmProjects\Brawl-Stars-AI\outputs\agc_revisit\2024-03-24_22-09-00\checkpoints\dataset"),
         Path("checkpoints\dataset"), dirs_exist_ok=True, )
     for f in Path("checkpoints\dataset").glob("*"):
-        if int(f.stem) > cfg.training.actor_critic.batch_num_samples:
+        if f.stem.isnumeric() and int(f.stem) > cfg.training.actor_critic.batch_num_samples:
             f.unlink()
     trainer.train_dataset.load_disk_checkpoint(trainer.ckpt_dir / 'dataset')
 
     # shutil.copytree(
-    #     Path(r"C:\Users\Michael\PycharmProjects\Brawl-Stars-AI\outputs\prioritization_and_accumulation\2024-03-27_23-33-28\checkpoints"),
+    #     Path(r"C:\Users\Michael\PycharmProjects\Brawl-Stars-AI\outputs\faster_and_fix\2024-03-31_04-47-02\checkpoints"),
     #     Path("checkpoints"), dirs_exist_ok=True, ignore=shutil.ignore_patterns("checkpoint*"))
     #
     # trainer.load_checkpoint()
@@ -568,7 +628,7 @@ def main(cfg: DictConfig):
     optimizer = trainer.optimizer_actor_critic
     ac_trainer = ACTrainer(trainer.cfg, env, actor, optimizer, trainer.train_dataset)
     ac_trainer.lr_scheduler.last_epoch = start_epoch - 1
-
+    ac_trainer.replay_only = True
 
     def collect_one(do_replay=False, do_log=False, do_step=False, verbose=0):
         ac_trainer.reset(collection=not do_replay)
@@ -603,25 +663,47 @@ def main(cfg: DictConfig):
 
     while len(ac_trainer.dataset) < ac_trainer.batch_size:
         collect_one()
+    collection_event = Event()
+    log_pipe_here, log_pipe_there = Pipe()
 
-    for n_episode in range(start_epoch, MAX_EPISODES):
-        for _ in range(cfg.training.actor_critic.collection_steps-1):
-            collect_one(do_replay=True, do_step=False, do_log=True, verbose=10)
+    def _wandb_log_listener():
+        while True:
+            if log_pipe_here.poll(1):
+                wandb.log(log_pipe_here.recv())
 
-        ac_trainer.replay_only = True
-        for _ in range(cfg.training.actor_critic.grad_acc_steps-cfg.training.actor_critic.collection_steps):
+    wandb_log_listener = Thread(target=_wandb_log_listener)
+    wandb_log_listener.start()
+
+    collection_process = Process(target=process_collection, args=(collection_event, cfg, env,
+                                                                  actor.to('cpu'),  # cuda weights pickling issue
+                                                                  ac_trainer.dataset,
+                                                                  Path("checkpoints/last.pt"),
+                                                                  log_pipe_there))
+    collection_process.start()
+    actor.to(device)
+    trainer.save_checkpoint(start_epoch, False, flush=False)
+    try:
+        for n_episode in range(start_epoch, MAX_EPISODES):
+            collection_event.set()
+            for _ in range(cfg.training.actor_critic.grad_acc_steps - 1):
+                ac_trainer.reset()
+                episode_metrics = ac_trainer.episode_end(do_step=False)
+                wandb.log({"epoch": n_episode, **episode_metrics})
             ac_trainer.reset()
-            episode_metrics = ac_trainer.episode_end(do_step=False)
+            episode_metrics = ac_trainer.episode_end(do_step=True)
             wandb.log({"epoch": n_episode, **episode_metrics})
-        ac_trainer.replay_only = False
 
-        collect_one(do_replay=True, do_step=True, do_log=True, verbose=10)  # ensure 1/1 step config works
+            while collection_event.is_set():  # wait for collection to finish
+                time.sleep(1)
 
-        print('Saving checkpoint')
-        if (n_episode + 1) % 50 == 0:
-            shutil.copytree('checkpoints', f'checkpoints/checkpoint_{n_episode}',
-                            ignore=shutil.ignore_patterns("dataset*", "checkpoint*"))
-        trainer.save_checkpoint(n_episode, False, flush=False)
+            print('Saving checkpoint')
+            if (n_episode + 1) % 50 == 0:
+                shutil.copytree('checkpoints', f'checkpoints/checkpoint_{n_episode}',
+                                ignore=shutil.ignore_patterns("dataset*", "checkpoint*"))
+            trainer.save_checkpoint(n_episode, True)
+    finally:
+        collection_process.terminate()
+        wandb.finish()
 
 
 @hydra.main(config_path=r"C:\Users\Michael\PycharmProjects\Brawl_iris\config", config_name="trainer")
@@ -629,11 +711,11 @@ def main_replay(cfg: DictConfig):
     trainer = Trainer(cfg)
     shutil.copytree(
         Path(
-            r"C:\Users\Michael\PycharmProjects\Brawl-Stars-AI\outputs\agc_revisit\2024-03-24_22-09-00\checkpoints\dataset"),
+            r"C:\Users\Michael\PycharmProjects\Brawl-Stars-AI\outputs\parallel_collection\2024-03-30_11-03-19\checkpoints\dataset"),
         Path("checkpoints\dataset"), dirs_exist_ok=True)
-    for f in Path("checkpoints\dataset").glob("*"):
-        if int(f.stem) > 48:
-            f.unlink()
+    # for f in Path("checkpoints\dataset").glob("*"):
+    #     if f.stem.isnumeric() and int(f.stem) > 48:
+    #         f.unlink()
     trainer.train_dataset.load_disk_checkpoint(trainer.ckpt_dir / 'dataset')
 
     start_epoch = trainer.start_epoch
@@ -649,11 +731,12 @@ def main_replay(cfg: DictConfig):
     ac_trainer.lr_scheduler.last_epoch = start_epoch - 1
 
     for n_step in tqdm.tqdm(range(start_epoch, MAX_EPISODES), desc="Step: ", total=MAX_EPISODES, ):
-        ac_trainer.reset(prioritize=False, zero_center=True)
+        ac_trainer.reset(prioritize=False)
         episode_metrics = ac_trainer.episode_end(epoch=n_step)
         print('Logging')
         wandb.log({"epoch": n_step, **episode_metrics})
         if n_step % 100 == 0:
+            continue
             print('Saving checkpoint')
             trainer.save_checkpoint(n_step, False, save_dataset=False)
 
@@ -661,11 +744,12 @@ def main_replay(cfg: DictConfig):
 @hydra.main(config_path=r"C:\Users\Michael\PycharmProjects\Brawl_iris\config", config_name="trainer")
 def inspect(cfg: DictConfig):
     try:
+        cfg.wandb.mode = "offline"
         trainer = Trainer(cfg)
 
         shutil.copytree(
             Path(
-                r"C:\Users\Michael\PycharmProjects\Brawl-Stars-AI\outputs\entropy_fix\2024-03-26_22-08-34\checkpoints"),
+                r"C:\Users\Michael\PycharmProjects\Brawl-Stars-AI\outputs\parallel_collection\2024-03-29_22-57-56\checkpoints"),
             Path("checkpoints"), dirs_exist_ok=True, ignore=shutil.ignore_patterns("checkpoint*"))
         trainer.load_checkpoint()
 
@@ -690,7 +774,7 @@ def inspect(cfg: DictConfig):
         ac_trainer.episode_end()
     finally:
         wandb.finish()
-        shutil.rmtree(os.getcwd())
+        shutil.rmtree(os.getcwd(), ignore_errors=True)
 
 
 if __name__ == "__main__":
